apps/polymarket/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ultramar-polymarket-backend"
version = "0.1.0"
description = "Backend services for the Polymarket vs Black-Scholes MVP"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "UNLICENSED" }
authors = [{ name = "Diegolden" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "httpx>=0.27",
  "websockets>=12.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.5", "pytest>=8.0", "respx>=0.21"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
